sys: bootinfo: Don't cast module addresses to 64 bits
%p expects a 32 bit (4 byte) address in IA32, and modules indeed have 32
bit addresses in the MultibootModule struct, so no reason in providing a
64 bit argument at all.

+++ sys/init/bootinfo.c
@@ -52,7 +52,7 @@ void RegisterBootModules(void) {
                 boot_modules[i]           = mods_arr[i];
                 LogMessage(LOG_DEBUG,
                            "Boot module %d: Start address %p, End Address %p, Command Line %s", i,
-                           (u64)boot_modules[i].start, (u64)boot_modules[i].end,
+                           boot_modules[i].start, boot_modules[i].end,
                            (char *)boot_modules[i].cmdline);
         }
 }
